Update revsliser import
I have modified the import of revolution slider slightly, sometime theme author need multiple slider to import for a single demo,  we are not able to do current code. So i check if the array index of current demo is array then it will import multiple demo.
Hope it will help many author to resolve this issue like me.